Abstracts are now formatted correctly.
When jekyll reads posts from the _posts directory, it seems
to insert <br/> tags for each newline.  This is a bit irritating,
not sure how to turn it off.  Fixed by moving the abstract
content to the YAML header, which doesn't seem to have this
limitation.  Probably is a more reasonable solution, in any case.